create methods for specifically validating birth and death declaration attachments
