+++ packages/gateway/src/features/correction/root-resolvers.ts
@@ -10,30 +10,23 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import { IAuthHeader } from '@gateway/common-types'
-import { GQLAttachmentInput, GQLResolver } from '@gateway/graphql/schema'
+import { GQLResolver } from '@gateway/graphql/schema'
 import { hasScope } from '@gateway/features/user/utils'
 import { buildFHIRBundle } from '@gateway/features/registration/fhir-builders'
 import { EVENT_TYPE } from '@gateway/features/fhir/constants'
 import { fetchFHIR, getIDFromResponse } from '@gateway/features/fhir/utils'
-import { validateAttachments } from '@gateway/utils/validators'
+import {
+  validateBirthDeclarationAttachments,
+  validateDeathDeclarationAttachments
+} from '@gateway/utils/validators'
 import { UserInputError } from 'apollo-server-hapi'
 
 export const resolvers: GQLResolver = {
   Mutation: {
     async requestBirthRegistrationCorrection(_, { id, details }, authHeader) {
       if (hasScope(authHeader, 'register')) {
-        const attachments = [
-          details.registration?.attachments,
-          details.informant?.affidavit,
-          details.mother?.photo,
-          details.father?.photo,
-          details.child?.photo
-        ]
-          .flat()
-          .filter((x): x is GQLAttachmentInput => x !== undefined)
-
         try {
-          await validateAttachments(attachments)
+          await validateBirthDeclarationAttachments(details)
         } catch (error) {
           throw new UserInputError(error.message)
         }
@@ -49,19 +42,8 @@ export const resolvers: GQLResolver = {
     },
     async requestDeathRegistrationCorrection(_, { id, details }, authHeader) {
       if (hasScope(authHeader, 'register')) {
-        const attachments = [
-          details.registration?.attachments,
-          details.informant?.affidavit,
-          details.mother?.photo,
-          details.father?.photo,
-          details.deceased?.photo,
-          details.spouse?.photo
-        ]
-          .flat()
-          .filter((x): x is GQLAttachmentInput => x !== undefined)
-
         try {
-          await validateAttachments(attachments)
+          await validateDeathDeclarationAttachments(details)
         } catch (error) {
           throw new UserInputError(error.message)
         }
